Added basic framework for triggering sound effects.
Added 'Beacon' from player- When the player presses the left mouse button, they will create a pulse with range '5', implanting an object of type 'BeaconInfo' in the memories of all nearby agents at key "LastBeacon". This object has the functions:

    GetTime() : float
returning the time the beacon was sent. This can be compared with Time.time

    GetPos() : Vector2
returning the position of the player's Legs object in the world.

This can be used to implement more advanced logic in sheep+wolves.